# swagger/swagger.yml
basePath: /api/v1
definitions:
...
  inventory.ServiceInput:
    properties:
      service_code:
        type: string
      service_name:
        type: string
      unit_price:
        type: number
    required:
      - service_code
      - service_name
      - unit_price
    type: object
  inventory.ServiceOutput:
    properties:
      id:
        type: integer
      service_code:
        type: string
      service_name:
        type: string
      unit_price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: hendri@kewr-digital.com
    name: Hendri Sulistya
  description: This is the API documentation for Kewr-ERP.
  title: Kewr-ERP API
  version: "1.0"
paths:
...
  /services:
    get:
      description: Retrieve all services from the database
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/inventory.ServiceOutput"
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Get all services
      tags:
        - Services
    post:
      consumes:
        - application/json
      description: Add a new service to the database
      parameters:
        - description: Service
          in: body
          name: service
          required: true
          schema:
            $ref: "#/definitions/inventory.ServiceInput"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/inventory.ServiceOutput"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Create a new service
      tags:
        - Services
  /services/{id}:
    delete:
      description: Delete a service from the database by its ID
      parameters:
        - description: Service ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Delete a service
      tags:
        - Services
    get:
      description: Retrieve a service from the database by its ID
      parameters:
        - description: Service ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/inventory.ServiceOutput"
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Get a service by ID
      tags:
        - Services
    put:
      consumes:
        - application/json
      description: Update a service in the database
      parameters:
        - description: Service ID
          in: path
          name: id
          required: true
          type: integer
        - description: Service
          in: body
          name: service
          required: true
          schema:
            $ref: "#/definitions/inventory.ServiceInput"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/inventory.ServiceOutput"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Update a service
      tags:
        - Services
...
schemes:
  - http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"